avoiding multiple initial writting on the db

+++ scripts/init.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
 # Espera a que la base de datos esté lista antes de aplicar migraciones y cargar datos
+echo " "
+
+echo "==================================================="
+
 echo "Esperando a que la base de datos esté disponible..."
 
 # Reintenta la conexión hasta que MySQL esté disponible
@@ -10,25 +14,37 @@ do
   sleep 5
 done
 
-echo "MySQL está listo. Generando data.json desde la base de datos SQLite."
+echo "MySQL está listo."
 
-# Genera el archivo data.json usando settings.py (que está configurado con SQLite)
-python manage.py dumpdata --settings=SoftServeAcademy.settings > /app/scripts/data.json
-echo "Archivo data.json generado con éxito desde SQLite."
+# Verifica si el archivo data.json no existe ya en el contenedor para evitar múltiples cargas
+if [ ! -f "/app/scripts/data.json" ]; then
+  echo "Generando data.json desde la base de datos SQLite."
+  # Genera el archivo data.json usando settings.py (configurado con SQLite)
+  python manage.py dumpdata --settings=SoftServeAcademy.settings > /app/scripts/data.json
+  echo "Archivo data.json generado con éxito desde SQLite."
+else
+  echo "Archivo data.json ya existe, omitiendo generación."
+fi
 
 # Cambia a settings_deployment (con MySQL) y ejecuta las migraciones
 echo "Aplicando migraciones en la base de datos MySQL..."
 python manage.py migrate --settings=SoftServeAcademy.settings_deployment
 echo "Migraciones aplicadas con éxito."
 
-# Carga los datos desde el archivo JSON recién generado en MySQL
-echo "Cargando datos desde data.json en la base de datos MySQL..."
-python manage.py loaddata /app/scripts/data.json --settings=SoftServeAcademy.settings_deployment
-echo "Datos cargados con éxito desde data.json."
+# Carga los datos en MySQL solo si la tabla de usuarios no existe
+if ! python manage.py inspectdb --settings=SoftServeAcademy.settings_deployment | grep -q "auth_user"; then
+  echo "Base de datos vacía, cargando datos desde data.json."
+  python manage.py loaddata /app/scripts/data.json --settings=SoftServeAcademy.settings_deployment
+  echo "Datos cargados con éxito desde data.json."
+else
+  echo "Base de datos ya inicializada, no se cargan datos."
+fi
 
-# Elimina el archivo data.json para que no permanezca en el contenedor
-rm /app/scripts/data.json
-echo "Archivo data.json eliminado del contenedor."
+# Elimina el archivo data.json si fue cargado
+if [ -f "/app/scripts/data.json" ]; then
+  rm /app/scripts/data.json
+  echo "Archivo data.json eliminado del contenedor."
+fi
 
 # Inicia el servidor de Django usando settings_deployment (con MySQL)
 echo "Iniciando el servidor de Django..."
